Don't dedup ConvParameters across different GPUs.
Previously, we assumed it was safe to autotune on GPU 0 and then use that
autotuning result on GPU 1, if the GPUs are the "same model".

It turns out this is not safe with the new cudnn 8 frontend API.

With the legacy API, a convolution autotuning result is just a number (or a
collection of a few numbers), so it's indeed safe to reuse this across GPUs.

But with the new API, an autotuning result is an in-memory execution plan
that's tied to a specific GPU.  You can't reuse it across different GPUs.

This change means that if you have two different GPUs, we will autotune the
same conv on both of them.  We were trying to avoid this, but oh well.

The serialization / deserialization code is not in yet, but when it is, we will
have to

 - resolve conflicts when serializing, if two different GPUs of identical model
   have different autotuning results.  Maybe just pick the first one.

 - populate autotuning results for each GPU with the same device kind string
   when deserializing.

PiperOrigin-RevId: 388336966
Change-Id: If357696fe5118a899fc6fd05ab96ce1130fcfb65