Fix overflow/crash in tf.range when limits is large
This PR tries to address the issue raised in 46913 where
tf.range (and implicitly tf.keras.layers.RepeatVector)
will overflow/crash when limits is large.

The reason of the overflow is that while calculating
the size within the kernel, the conditional statements
comes with `int64 = cond ? int64 : double` will implicitly
convert to double first and then cast back to int64, causing
the overflow and crash.

This PR fixes the issue by casting to int64 in both selections
within the conditional statements first.

This PR fixes 46913.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>