Add TF_ENABLED_DEVICE_TYPES env variable to control which devices TF enables.
Many JAX users make heavy use of TensorFlow on CPU for tasks such as data
loading, processing, saved model etc. These users will use JAX to run programs
on GPU/TPU.

GPU users have reported running near the limit of GPU memory, and we have
tracked this down to them unknowingly running some operations on GPU with TF,
and TF allocating a significant chunk of GPU memory on top of JAX's chunk. This
happens particularly often in TF2 (some statements (e.g. `tfds.load(..)`)
eagerly create TF constants placed on the GPU, users struggle to identify this
code as being responsible for using TF on GPU).

It would be convenient if there was an environment variable we could set to
force all usages of TF in a given program to be CPU only. This is similar in
spirit to variables such as CUDA_VISIBLE_DEVICES. It would be easy for training
framework developers to deploy for all of their users (e.g. in the same way we
set some sensible defaults for NCCL to improve performance). Given this CL we
can implement this by setting `TF_ENABLED_DEVICE_TYPES=CPU`.

PiperOrigin-RevId: 389741331
Change-Id: Iad30b9eb96ff720f942b0f64f654da9d0555f2a3