[TF:TRT] Fix a problem in running a TRTEngineOp native segment with int32 values as inputs.
A TRTEngineOp has all its inputs on device memory. On the other hand,
TensorFlow assumes all function arguments with int32 values are on host memory
unless the kIntsOnDeviceAttr attribute is set to true. This change sets up the
kIntsOnDeviceAttr attribute for the TRTEngineOp native segment function to
support the TensorFlow execution of the function.

Modify an existing test to test the situation.

PiperOrigin-RevId: 389889656
Change-Id: Iea07647da8c233f97aebb1de0c31fa561b0e3556