[XLA:GPU] Simplify heuristics for promotion to int8x32.
Previously, we would upgrade to int8x32 if doing so did not result in a > 2.5x
increase in the input/filter/output size of the conv.

Now we only consider the input/output sizes.  Sometimes the filter increases by
more than 2.5x and we observe that it's still worthwhile to pad.  Anyway I
think this makes more sense, since the increase in the filter size is really a
function of the input/output increases.

As a result we're able to switch the heuristic to the more intuitive "don't
expand input/output by 2x or more".

PiperOrigin-RevId: 390144519
Change-Id: I96cdb0a95dc627157cdcd02d4eb1c593cedd148f