Remove accidentally added Python symbols.
The following symbols were accidentally added. This change removes these symbols.
  * tf.quantize_and_dequantize_v4
  * tf.batch_mat_mul_v3
  * tf.sparse_segment_sum_grad

Each of the three changes that accidentally added a symbol did so by creating a new op, then forgetting to create a Python ApiDef marking the op as HIDDEN. The ops were not intended to be directly exposed under the top-level tf namespace.

There are no GitHub usages of any of these symbols, so removing them is safe.

PiperOrigin-RevId: 390245011
Change-Id: I2f91704a431453f136b304711980bc92e4c2c214