Adopt signature concept in representative_datagen
As TFLite team delivered the Signature concept recently
(https://www.tensorflow.org/lite/guide/signatures), it is a good time to rethink
about the representative_dataset method behavior.

As per the TensorFlow product design and TensorFlow MLIR behaviors, it is
really impossible to keep the input ordering same as before.

Instead of depending on the tensor indices, it would be better to depend on the
input names, specified in the signature table.

For example,

def representative_dataset():
    for _ in range(100):
      image = np.random.rand(1, 244, 244, 3).astype(np.float32)
      bias = np.random.rand(1, 10).astype(np.float32)
      yield {"image": image, "bias": bias}

PiperOrigin-RevId: 390323891
Change-Id: I191de757452bcf52c3fed6457713934ee9c4de52