update `@pasta` to work when `@org_tensorflow` is *not* the main repo
`@pasta`'s BUILD file uses a helper macro (in `build_defs.bzl`) that
lives in `@org_tensorflow` which it references using `@//` (the main
repo). This breaks when `@org_tensorflow` isn't the main repo; i.e. if
it's a dependency in another workspace.

This commit changes the setup for the `@pasta` repo to symlink
`build_defs.bzl` into the `@pasta` repo.