First step towards splitting xla_legalize_tf target.
The goal is to offer the PopulateXYPatterns in their own BUILD targets so that
they can be used selectively instead of just depending on all of them.
The xla_legalize_tf target will stay and offer the existing functionality of
pulling in all passes.
The legalize_tf target contains the main lowering patterns from TF dialect to HLO dialect.

PiperOrigin-RevId: 391465345
Change-Id: I26447429c1b333fe568826081720b02e1034e6ce