Fix breakage caused by missing implicit conversion.
In C++14 there is no implicit conversion from `llvm::StringRef&` to `absl::string_view` but it exists in C++17. As TF is still on C++14, LLVM/MLIR/TFRT updates that assume C++17 will break TF and will need similar fixes as this one.

Enabling C++17 dialect in TF is on a longer roadmap.

PiperOrigin-RevId: 392558945
Change-Id: Ia1d575ca96936fe4304e130631cff61ef26f6638