Enable mhlo.Iota fold and mhlo.BroadcastInDim for some case.
In particular, the mhlo.Iota fold supports the case when iota_dimension = last_dimension = output_rank-1. Additionally, the mhlo.BroadcastInDim fold supports the case where operand is a rank1 vector and the broadcast_dim = last_dim. We do not fold Iota or BroadCastInDim if the total number of elements is greater than 1024.

Also add a new case for MatchIota (folded 1d Iota + broadCastInDim).

PiperOrigin-RevId: 394399433
Change-Id: I5f56b214fc1786f64eb1210c5f6828aeb57b07bb