Skip get-tuple-element nodes in graphviz.
Re-land cl/394286090, which was rolled back in cl/394760138 due to
CHECK-failures.

The only difference between this and the original is inside GetNodeForEdge.  In
the original, we unnested fusions first, then unnested GTEs.  Now we reverse
the order; first we unnest GTEs, then we unnest fusions.  This is necessary to
avoid crashes when rendering multi-output-fusions.

The tuple index is written inside the consuming instruction as e.g.  "operand 1
= tuple-element 3 of fusion.42".

Alternative considered: My first attempt was to put the tuple element in the
edge label, which right now contains the operand index.  This didn't work well.
Graphviz doesn't prevent the edge label from overlapping with the edge itself
or anything else in the graph.  It was irredeemably messy.

PiperOrigin-RevId: 395251901
Change-Id: I39b1f9731fba1ed89f0e0330221d62cfc7bd0f6f