Update Dockerfiles to Ubuntu 20.04
This seemed to mostly work locally, so I'm making this change to see if
our automation will pick it up and succeed. This also updates the
container Python version to one newer than 3.6 and the CUDA version to
11.2.

One notable new requirement is that users must run ldconfig after the
image starts to pick up libraries that are mounted by nvidia-docker. If
the automation is successful, then I'll look into updating the Docker
Hub page.

Working with this system is getting to be quite the hassle, so I'm glad
to be able to drop it in the future.