Fix crash of max_pool3d when ksize is 0 or negative
This PR tries to address the issue raised in 51936 where
max_pool3d will crash when any dim of ksize is 0 or negative.

While the original issue was raised toward tf.keras.layers.MaxPooling3D,
the issue can also be triggered when max_pool3d is called directly
with tensorflow itself.

For that reason a separate fix inside tensorflow is also fixed here.

This PR fixes 51936.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>