Fix TFG func op parsing use-after-free (NFC)
In cases where the SmallVector containing the string for the function
operands would grow during parsing, it invalidates the addresses of
the strings that were previously captured as StringRef in the `entry_args`
array. We change the SmallVector of string to use unique_ptr<string>
instead: the indirection preserve the validity of the StringRef even
through vector growth.
PiperOrigin-RevId: 396416141
Change-Id: Iab3c19cdff8d76197eb6fa79d251ecf6a18fbf1b