NativeInterpreterWrapper refactoring: split applyDelegates.
This is a small refactoring that is a partial step towards the
goal of eventually replacing the use of the experimental C++
API method Interpreter::ModifyGraphWithDelegate with calls to
InterpreterOptions::AddDelegate.

This step splits the existing Java applyDelegates method into two,
addDelegates and applyDelegates, so we now add all the delegates
(except for the XNNPACK delegate) to the 'delegates' list before
constructing the interpreter; and also moves the place where the
delegates are applied closer to where the interpreter is constructed;
but we still use ModifyGraphWithDelegate to apply the delegates.

PiperOrigin-RevId: 396588873
Change-Id: Iee98f753f8f530438b7933215c92316592b4ba51