[MLIR][TF][SCF] Introduce tf.IfRegion -> scf.if lowering
Lowering for `tf.IfRegion` -> `scf.if` implemented in a new pass
`convert-tf-control-flow-to-scf`, which is added in
`tensorflow/transforms` and can be used for all direct control flow
lowerings from the TensorFlow dialect to the SCF dialect (without going
through dialects like HLO, which involve tuples, for example). Currently
this pass has one pattern `ConvertIfRegionOp`. It lowers the
`tf.IfRegion` op to `scf.if`.

Signed-off-by: Srishti Srivastava <srishti.srivastava@polymagelabs.com>