[retry]Fix a shape mismatch when partial batch handling is enabled
Partial batch handling will make the batch dimension unknown because of the
tf.cond(has_value, true_fn=get_value, false_fn=make_dummy). This tf.cond is
executed in DistributedIterator.get_next_as_optional() even when partial batch
handling is disabled. To counter this unnecessary shape lost, we recover the shape after tf.cond with the shape of the true branch. However we may recover it with the wrong shape when partial batch handling is enabled (as we could take the false branch).

This change refactors the partial batch handling logic so that
DistributedIterator.get_next() and DistributedIterator.get_next_as_optional()
is consistent.

The previous _get_next_as_optional is a misleading name as it does parial batch
handling and doesn't produce optional values. Hope that the picture is more
clear after this change.

PiperOrigin-RevId: 397253580
Change-Id: Ia09bc96c8bcda721d0d3add33f1b6540f37a4b48